+++ packages/vitepress-plugin-rss/src/utils/index.ts
@@ -1,6 +1,6 @@
 import os from 'node:os'
 import path from 'node:path'
-import { spawnSync } from 'node:child_process'
+import { spawn } from 'cross-spawn'
 
 export function formatDate(d: any, fmt = 'yyyy-MM-dd hh:mm:ss') {
   if (!(d instanceof Date)) {
@@ -44,73 +44,38 @@ export function normalizePath(id: string): string {
 }
 
 export function getDefaultTitle(content: string) {
-  const title
-    = clearMatterContent(content)
-      .split('\n')
-      ?.find((str) => {
-        return str.startsWith('# ')
-      })
-      ?.slice(2)
-      .replace(/[\s]/g, '') || ''
-  return title
+  const match = content.match(/^(#+)\s+(.+)/m)
+  return match?.[2] || ''
 }
 
-export function clearMatterContent(content: string) {
-  let first___: unknown
-  let second___: unknown
+const cache = new Map<string, Date>()
 
-  const lines = content.split('\n').reduce<string[]>((pre, line) => {
-    // 移除开头的空白行
-    if (!line.trim() && pre.length === 0) {
-      return pre
-    }
-    if (line.trim() === '---') {
-      if (first___ === undefined) {
-        first___ = pre.length
-      }
-      else if (second___ === undefined) {
-        second___ = pre.length
-      }
-    }
-    pre.push(line)
-    return pre
-  }, [])
-  return (
-    lines
-      // 剔除---之间的内容
-      .slice((second___ as number) || 0)
-      .join('\n')
-  )
-}
+export function getFileBirthTime(url: string): Promise<Date | undefined> | Date {
+  const cached = cache.get(url)
+  if (cached)
+    return cached
 
-export function getFileBirthTime(url: string) {
-  let date = new Date()
-
-  try {
-    // 参考 vitepress 中的 getGitTimestamp 实现
-    // const infoStr = execSync(`git log -1 --pretty="%ci" ${url}`)
-    //   .toString('utf-8')
-    //   .trim()
-
-    const infoStr = spawnSync('git', ['log', '-1', '--pretty="%ci"', url])
-      .stdout?.toString()
-      .replace(/["']/g, '')
-      .trim()
-    if (infoStr) {
-      date = new Date(infoStr)
-    }
-  }
-  catch (error) {
-    return formatDate(date)
-  }
-
-  return formatDate(date)
+  return new Promise((resolve) => {
+    // 使用异步回调
+    const child = spawn('git', ['log', '-1', '--pretty="%ai"', url])
+    let output = ''
+    child.stdout.on('data', d => (output += String(d)))
+    child.on('close', () => {
+      const date = new Date(output)
+      cache.set(url, date)
+      resolve(date)
+    })
+    child.on('error', () => {
+      resolve(undefined)
+    })
+  })
 }
 
 export function getTextSummary(text: string, count = 100) {
   return (
-    clearMatterContent(text)
-      .match(/^# ([\s\S]+)/m)?.[1]
+    text
+      // 首个标题
+      ?.replace(/^#+\s+.*/, '')
       // 除去标题
       ?.replace(/#/g, '')
       // 除去图片
@@ -121,24 +86,10 @@ export function getTextSummary(text: string, count = 100) {
       ?.replace(/\*\*(.*?)\*\*/g, '$1')
       ?.split('\n')
       ?.filter(v => !!v)
-      ?.slice(1)
       ?.join('\n')
       ?.replace(/>(.*)/, '')
+      ?.replace(/</g, '&lt;').replace(/>/g, '&gt;')
+      ?.trim()
       ?.slice(0, count)
   )
-}
-
-export const EXTERNAL_URL_RE = /^[a-z]+:/i
-
-/**
- * Join two paths by resolving the slash collision.
- */
-export function joinPath(base: string, path: string): string {
-  return `${base}${path}`.replace(/\/+/g, '/')
-}
-
-export function withBase(base: string, path: string) {
-  return EXTERNAL_URL_RE.test(path) || path.startsWith('.')
-    ? path
-    : joinPath(base, path)
 }