+++ packages/vitepress-plugin-rss/CHANGELOG.md
@@ -1,5 +1,36 @@
 # vitepress-plugin-rss
 
+## 0.2.6
+
+### Patch Changes
+
+- fix: dev pending and build failed
+
+## 0.2.5
+
+### Patch Changes
+
+- feat: Support for customized HTML generation logic
+
+## 0.2.4
+
+### Patch Changes
+
+- fix: generate title error
+
+## 0.2.3
+
+### Patch Changes
+
+- chore: add peerdeps
+- chore: clean code
+
+## 0.2.2
+
+### Patch Changes
+
+- feat: support `ignorePublish` option，include `publish: false` frontmatter docs
+
 ## 0.2.1
 
 ### Patch Changes